requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
langchain-community
python-dotenv
pypdf
faiss-cpu
langchain-deepseek
gradio==5.16.0